requirements.txt
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0
faiss-cpu>=1.7.4
streamlit>=1.36.0
pydantic>=2.0.0
pytest>=8.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
groq>=0.8.0
httpx>=0.27.0
sentence-transformers
